% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coef.linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{coef}
\usage{
\method{coef}{linreg}(result)
}
\arguments{
\item{what}{results of the use of linreg function}
}
\value{
a named vector of the estimate of the coefficients
}
\description{
this function only separates the estimate of the coefficients apart.
}

