% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.linreg.R
\name{plot}
\alias{plot}
\title{plot}
\arguments{
\item{what}{results of the use of linreg function}
}
\value{
the two plots described over
}
\description{
plots two graphs with the values calculated by linreg function :
 - one plots residuals vs fitted values
 - the other one plots the scale location, ie the square root of the
 absolute value of the residuals
}

