% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pred.linreg.R
\name{pred}
\alias{pred}
\title{pred}
\usage{
pred(result, ...)
}
\arguments{
\item{what}{results of the use of linreg function, and a numeric vector, which length is the
same as the number of coefficients.}
}
\value{
a vector with the predicted values
}
\description{
this function predicts the values of the parameter we want to explain,
ie it calculates y hat for the vector of values in parameter
}

