% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.ridgereg.R
\name{print ridge}
\alias{print ridge}
\alias{print.ridgereg}
\title{print ridge}
\usage{
\method{print}{ridgereg}(X)
}
\arguments{
\item{what}{results of the use of ridgereg function}
}
\value{
the call, ie the formula of the regression, and the estimate of the coefficients,
with their names
}
\description{
this function prints important results calculated by ridgereg function
}

