% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{print}
\usage{
\method{print}{linreg}(X)
}
\arguments{
\item{what}{results of the use of linreg function}
}
\value{
the call, ie the formula of the regression, and the estimate of the coefficients,
with their names
}
\description{
this function prints important results calculated by linreg function
}

