% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resid.linreg.R
\name{resid}
\alias{resid}
\title{resid}
\usage{
resid(X)
}
\arguments{
\item{what}{results of the use of linreg function}
}
\value{
a vector with the resiudals
}
\description{
this function takes apart the residuals of the linear regression
calculated by linreg function
}

