% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{data}
\name{ridgereg}
\alias{formula}
\alias{ridgereg}
\title{ridgereg}
\format{\preformatted{Class 'formula' length 3 Sepal.Length ~ Sepal.Width + Petal.Width
  ..- attr(*, ".Environment")=<environment: namespace:lab4bis> 
}}
\usage{
formula
}
\arguments{
\item{lambda}{- hyperparameter to be tuned. if not specified, we set it to 0.}

\item{a}{"formula" of the form y ~ x+... and a dataframe}
}
\description{
Uses ridge regression to fit a model.
}
\author{
Niclas Lovsjö & Maxime Bonneau
}
\references{
ESLII: http://web.stanford.edu/~hastie/local.ftp/Springer/OLD/ESLII_print4.pdf
}
\keyword{datasets}

