% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{visualize_airport_delays}
\arguments{
\item{no}{parameters}
}
\description{
Plots the mean delay of ﬂights for different airports by longitude and latitude using ggplot2
}
\details{

}
\author{
Niclas Lovsjö & Maxime Bonneau
}
\references{
https://www.rstudio.com/wp-content/uploads/2015/02/data-wrangling-cheatsheet.pdf
https://cran.r-project.org/web/packages/nycflights13/nycflights13.pdf
}

